remove databaseType
