--- src/main/java/com/github/wz2cool/dynamic/mybatis/DatabaseType.java
@@ -1,8 +1,0 @@
-package com.github.wz2cool.dynamic.mybatis;
-
-/**
- * Created by Frank on 7/17/2017.
- */
-public enum DatabaseType {
-    MYSQL, SQLSERVER, H2, POSTRESQL
-}
