--- src/main/java/com/github/wz2cool/dynamic/mybatis/mapper/DynamicQueryMapper.java
@@ -1,15 +1,0 @@
-package com.github.wz2cool.dynamic.mybatis.mapper;
-
-import tk.mybatis.mapper.common.BaseMapper;
-
-/**
- * \* Created with IntelliJ IDEA.
- * \* User: Frank
- * \* Date: 8/7/2017
- * \* Time: 1:53 PM
- * \* To change this template use File | Settings | File Templates.
- * \* Description:
- * \
- */
-public interface DynamicQueryMapper<T> extends BaseMapper<T> {
-}
