--- src/main/java/com/github/wz2cool/dynamic/mybatis/mapper/SelectByDynamicQueryMapper.java
@@ -1,13 +1,0 @@
-package com.github.wz2cool.dynamic.mybatis.mapper;
-
-/**
- * \* Created with IntelliJ IDEA.
- * \* User: Frank
- * \* Date: 8/7/2017
- * \* Time: 1:54 PM
- * \* To change this template use File | Settings | File Templates.
- * \* Description:
- * \
- */
-public interface SelectByDynamicQueryMapper<T> {
-}
