+++ src/test/java/com/github/wz2cool/dynamic/mybatis/db/mapper/UserDao.java
@@ -1,7 +1,7 @@
 package com.github.wz2cool.dynamic.mybatis.db.mapper;
 
 import com.github.wz2cool.dynamic.mybatis.db.model.entity.table.User;
-import com.github.wz2cool.dynamic.mybatis.mapper.DynamicQueryMapper;
+import tk.mybatis.mapper.common.Mapper;
 
 /**
  * \* Created with IntelliJ IDEA.
@@ -12,5 +12,5 @@ import com.github.wz2cool.dynamic.mybatis.mapper.DynamicQueryMapper;
  * \* Description:
  * \
  */
-public interface UserDao extends DynamicQueryMapper<User> {
+public interface UserDao extends Mapper<User> {
 }
